docker-compose: run tests from a real client
There's no need to run the test suite from inside the 'NFS server host',
i.e., the 'pod' container network (using traffic over 127.0.0.1).

Instead, run the test suite in a separate container, also contained to
the network, and connect to all services 'from the outside', as it would
be for real (and the way the Kubernetes tests work).